microservices/location/requirements.txt
psycopg2
python-dotenv
fastapi